refactor(#8/post): validator 패키지 위치 변경
