+++ src/main/java/org/sopt/domain/post/application/validator/PostValidationRule.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.post.validator;
+package org.sopt.domain.post.application.validator;
 
 public interface PostValidationRule {
     void validate(Long userId, String title);
