+++ src/main/java/org/sopt/domain/post/application/validator/rule/DuplicateTitleValidator.java
@@ -1,5 +1,6 @@
-package org.sopt.domain.post.validator;
+package org.sopt.domain.post.application.validator.rule;
 
+import org.sopt.domain.post.application.validator.PostValidationRule;
 import org.sopt.global.error.BusinessException;
 import org.sopt.domain.post.presentation.exception.PostErrorCode;
 import org.sopt.domain.post.infrastructure.repository.PostRepository;
